feat: relocate tests in test directory per module
