+++ src/user/__test__/components/UserTitle.spec.ts
@@ -1,5 +1,5 @@
 import { render } from '@testing-library/vue';
-import UserTitle from './UserTitle.vue';
+import UserTitle from '../../components/UserTitle.vue';
 
 describe('[Component] Given UserTitle model', () => {
